Renamed camelCase filenames to use kebab case
